refactor(mongo_client): move floating functions to mongo_client_ops
Fixes NODE-1255
