+++ lib/mongo_client.js
@@ -1,29 +1,17 @@
 'use strict';
 
-const parse = require('mongodb-core').parseConnectionString;
-const Server = require('./topologies/server');
-const Mongos = require('./topologies/mongos');
-const ReplSet = require('./topologies/replset');
-const EventEmitter = require('events').EventEmitter;
-const inherits = require('util').inherits;
-const deprecate = require('util').deprecate;
-const ReadPreference = require('mongodb-core').ReadPreference;
-const Logger = require('mongodb-core').Logger;
-const MongoError = require('mongodb-core').MongoError;
-const handleCallback = require('./utils').handleCallback;
-const Db = require('./db');
-const f = require('util').format;
-const shallowClone = require('./utils').shallowClone;
-const authenticate = require('./authenticate');
-const ServerSessionPool = require('mongodb-core').Sessions.ServerSessionPool;
-const executeOperation = require('./utils').executeOperation;
 const ChangeStream = require('./change_stream');
+const Db = require('./db');
+const EventEmitter = require('events').EventEmitter;
+const executeOperation = require('./utils').executeOperation;
+const handleCallback = require('./utils').handleCallback;
+const inherits = require('util').inherits;
+const MongoError = require('mongodb-core').MongoError;
 
-const legacyParse = deprecate(
-  require('./url_parser'),
-  'current URL string parser is deprecated, and will be removed in a future version. ' +
-    'To use the new parser, pass option { useNewUrlParser: true } to MongoClient.connect.'
-);
+// Operations
+const connectOp = require('./operations/mongo_client_ops').connectOp;
+const logout = require('./operations/mongo_client_ops').logout;
+const validOptions = require('./operations/mongo_client_ops').validOptions;
 
 /**
  * @fileOverview The **MongoClient** class is a class that allows for making Connections to MongoDB.
@@ -57,140 +45,6 @@ const legacyParse = deprecate(
  *   client.close();
  * });
  */
-var validOptionNames = [
-  'poolSize',
-  'ssl',
-  'sslValidate',
-  'sslCA',
-  'sslCert',
-  'sslKey',
-  'sslPass',
-  'sslCRL',
-  'autoReconnect',
-  'noDelay',
-  'keepAlive',
-  'keepAliveInitialDelay',
-  'connectTimeoutMS',
-  'family',
-  'socketTimeoutMS',
-  'reconnectTries',
-  'reconnectInterval',
-  'ha',
-  'haInterval',
-  'replicaSet',
-  'secondaryAcceptableLatencyMS',
-  'acceptableLatencyMS',
-  'connectWithNoPrimary',
-  'authSource',
-  'w',
-  'wtimeout',
-  'j',
-  'forceServerObjectId',
-  'serializeFunctions',
-  'ignoreUndefined',
-  'raw',
-  'bufferMaxEntries',
-  'readPreference',
-  'pkFactory',
-  'promiseLibrary',
-  'readConcern',
-  'maxStalenessSeconds',
-  'loggerLevel',
-  'logger',
-  'promoteValues',
-  'promoteBuffers',
-  'promoteLongs',
-  'domainsEnabled',
-  'checkServerIdentity',
-  'validateOptions',
-  'appname',
-  'auth',
-  'user',
-  'password',
-  'authMechanism',
-  'compression',
-  'fsync',
-  'readPreferenceTags',
-  'numberOfRetries',
-  'auto_reconnect',
-  'minSize',
-  'monitorCommands',
-  'retryWrites',
-  'useNewUrlParser'
-];
-
-const validOptionsLowerCaseToCamelCase = validOptionNames.reduce((obj, name) => {
-  obj[name.toLowerCase()] = name;
-  return obj;
-}, {});
-
-function transformUrlOptions(_object) {
-  let object = Object.assign({ servers: _object.hosts }, _object.options);
-  for (let name in object) {
-    const camelCaseName = validOptionsLowerCaseToCamelCase[name];
-    if (camelCaseName) {
-      object[camelCaseName] = object[name];
-    }
-  }
-  if (_object.auth) {
-    const auth = _object.auth;
-    for (let i in auth) {
-      if (auth[i]) {
-        object[i] = auth[i];
-      }
-    }
-    if (auth.username) {
-      object.auth = auth;
-      object.user = auth.username;
-    }
-    if (auth.db) {
-      object.dbName = auth.db;
-    }
-  }
-  if (object.maxpoolsize) {
-    object.poolSize = object.maxpoolsize;
-  }
-  if (object.readconcernlevel) {
-    object.readConcern = { level: object.readconcernlevel };
-  }
-  if (object.wtimeoutms) {
-    object.wtimeout = object.wtimeoutms;
-  }
-  return object;
-}
-
-function legacyTransformUrlOptions(object) {
-  return mergeOptions(createUnifiedOptions({}, object), object, false);
-}
-
-var ignoreOptionNames = ['native_parser'];
-var legacyOptionNames = ['server', 'replset', 'replSet', 'mongos', 'db'];
-
-function validOptions(options) {
-  var _validOptions = validOptionNames.concat(legacyOptionNames);
-
-  for (var name in options) {
-    if (ignoreOptionNames.indexOf(name) !== -1) {
-      continue;
-    }
-
-    if (_validOptions.indexOf(name) === -1 && options.validateOptions) {
-      return new MongoError(f('option %s is not supported', name));
-    } else if (_validOptions.indexOf(name) === -1) {
-      console.warn(f('the options [%s] is not supported', name));
-    }
-
-    if (legacyOptionNames.indexOf(name) !== -1) {
-      console.warn(
-        f(
-          'the server/replset/mongos options are deprecated, ' +
-            'all their options are supported at the top level of the options object [%s]',
-          validOptionNames
-        )
-      );
-    }
-  }
-}
 
 /**
  * Creates a new MongoClient instance
@@ -275,7 +129,7 @@ function MongoClient(url, options) {
   };
 
   // Get the promiseLibrary
-  var promiseLibrary = this.s.options.promiseLibrary || Promise;
+  const promiseLibrary = this.s.options.promiseLibrary || Promise;
 
   // Add the promise to the internal state
   this.s.promiseLibrary = promiseLibrary;
@@ -306,7 +160,7 @@ inherits(MongoClient, EventEmitter);
  */
 MongoClient.prototype.connect = function(callback) {
   // Validate options object
-  var err = validOptions(this.s.options);
+  const err = validOptions(this.s.options);
 
   if (typeof callback === 'string') {
     throw new TypeError('`connect` only accepts a callback');
@@ -314,16 +168,6 @@ MongoClient.prototype.connect = function(callback) {
 
   return executeOperation(this, connectOp, [this, err, callback], {
     skipSessions: true
-  });
-};
-
-const connectOp = (self, err, callback) => {
-  // Did we have a validation error
-  if (err) return callback(err);
-  // Fallback to callback based connect
-  connect(self, self.s.url, self.s.options, function(err) {
-    if (err) return callback(err);
-    callback(null, self);
   });
 };
 
@@ -340,17 +184,10 @@ MongoClient.prototype.logout = function(options, callback) {
   options = options || {};
 
   // Establish the correct database name
-  var dbName = this.s.options.authSource ? this.s.options.authSource : this.s.options.dbName;
+  const dbName = this.s.options.authSource ? this.s.options.authSource : this.s.options.dbName;
 
   return executeOperation(this, logout, [this, dbName, callback], {
     skipSessions: true
-  });
-};
-
-const logout = (self, dbName, callback) => {
-  self.topology.logout(dbName, function(err) {
-    if (err) return callback(err);
-    callback(null, true);
   });
 };
 
@@ -362,31 +199,30 @@ const logout = (self, dbName, callback) => {
  * @return {Promise} returns Promise if no callback passed
  */
 MongoClient.prototype.close = function(force, callback) {
-  var self = this;
   if (typeof force === 'function') (callback = force), (force = false);
 
   // Close the topology connection
   this.topology.close(force);
 
   // Emit close event
-  self.emit('close', self);
+  this.emit('close', this);
 
   // Fire close event on any cached db instances
-  for (var name in this.s.dbCache) {
+  for (const name in this.s.dbCache) {
     this.s.dbCache[name].emit('close');
   }
 
   // Remove listeners after emit
-  self.removeAllListeners('close');
+  this.removeAllListeners('close');
 
   // Callback after next event loop tick
   if (typeof callback === 'function')
-    return process.nextTick(function() {
+    return process.nextTick(() => {
       handleCallback(callback, null);
     });
 
   // Return dummy promise
-  return new this.s.promiseLibrary(function(resolve) {
+  return new this.s.promiseLibrary(resolve => {
     resolve();
   });
 };
@@ -413,7 +249,7 @@ MongoClient.prototype.db = function(dbName, options) {
   }
 
   // Copy the options and add out internal override of the not shared flag
-  var finalOptions = Object.assign({}, this.s.options, options);
+  const finalOptions = Object.assign({}, this.s.options, options);
 
   // Do we have the db in the cache already
   if (this.s.dbCache[dbName] && finalOptions.returnNonCachedInstance !== true) {
@@ -429,7 +265,7 @@ MongoClient.prototype.db = function(dbName, options) {
   }
 
   // Return the db object
-  var db = new Db(dbName, this.topology, finalOptions);
+  const db = new Db(dbName, this.topology, finalOptions);
 
   // Add the db to the cache
   this.s.dbCache[dbName] = db;
@@ -525,13 +361,13 @@ MongoClient.prototype.isConnected = function(options) {
  * @return {Promise<MongoClient>} returns Promise if no callback passed
  */
 MongoClient.connect = function(url, options, callback) {
-  var args = Array.prototype.slice.call(arguments, 1);
+  const args = Array.prototype.slice.call(arguments, 1);
   callback = typeof args[args.length - 1] === 'function' ? args.pop() : undefined;
   options = args.length ? args.shift() : null;
   options = options || {};
 
   // Create client
-  var mongoClient = new MongoClient(url, options);
+  const mongoClient = new MongoClient(url, options);
   // Execute the connect method
   return mongoClient.connect(callback);
 };
@@ -625,456 +461,6 @@ MongoClient.prototype.watch = function(pipeline, options) {
   }
 
   return new ChangeStream(this, pipeline, options);
-};
-
-var mergeOptions = function(target, source, flatten) {
-  for (var name in source) {
-    if (source[name] && typeof source[name] === 'object' && flatten) {
-      target = mergeOptions(target, source[name], flatten);
-    } else {
-      target[name] = source[name];
-    }
-  }
-
-  return target;
-};
-
-var createUnifiedOptions = function(finalOptions, options) {
-  var childOptions = [
-    'mongos',
-    'server',
-    'db',
-    'replset',
-    'db_options',
-    'server_options',
-    'rs_options',
-    'mongos_options'
-  ];
-  var noMerge = ['readconcern', 'compression'];
-
-  for (var name in options) {
-    if (noMerge.indexOf(name.toLowerCase()) !== -1) {
-      finalOptions[name] = options[name];
-    } else if (childOptions.indexOf(name.toLowerCase()) !== -1) {
-      finalOptions = mergeOptions(finalOptions, options[name], false);
-    } else {
-      if (
-        options[name] &&
-        typeof options[name] === 'object' &&
-        !Buffer.isBuffer(options[name]) &&
-        !Array.isArray(options[name])
-      ) {
-        finalOptions = mergeOptions(finalOptions, options[name], true);
-      } else {
-        finalOptions[name] = options[name];
-      }
-    }
-  }
-
-  return finalOptions;
-};
-
-function translateOptions(options) {
-  // If we have a readPreference passed in by the db options
-  if (typeof options.readPreference === 'string' || typeof options.read_preference === 'string') {
-    options.readPreference = new ReadPreference(options.readPreference || options.read_preference);
-  }
-
-  // Do we have readPreference tags, add them
-  if (options.readPreference && (options.readPreferenceTags || options.read_preference_tags)) {
-    options.readPreference.tags = options.readPreferenceTags || options.read_preference_tags;
-  }
-
-  // Do we have maxStalenessSeconds
-  if (options.maxStalenessSeconds) {
-    options.readPreference.maxStalenessSeconds = options.maxStalenessSeconds;
-  }
-
-  // Set the socket and connection timeouts
-  if (options.socketTimeoutMS == null) options.socketTimeoutMS = 360000;
-  if (options.connectTimeoutMS == null) options.connectTimeoutMS = 30000;
-
-  // Create server instances
-  return options.servers.map(function(serverObj) {
-    return serverObj.domain_socket
-      ? new Server(serverObj.domain_socket, 27017, options)
-      : new Server(serverObj.host, serverObj.port, options);
-  });
-}
-
-var events = [
-  'timeout',
-  'close',
-  'serverOpening',
-  'serverDescriptionChanged',
-  'serverHeartbeatStarted',
-  'serverHeartbeatSucceeded',
-  'serverHeartbeatFailed',
-  'serverClosed',
-  'topologyOpening',
-  'topologyClosed',
-  'topologyDescriptionChanged',
-  'commandStarted',
-  'commandSucceeded',
-  'commandFailed',
-  'joined',
-  'left',
-  'ping',
-  'ha',
-  'all',
-  'fullsetup'
-];
-
-//
-// Collect all events in order from SDAM
-//
-function collectEvents(self, topology) {
-  var collectedEvents = [];
-
-  if (self instanceof MongoClient) {
-    events.forEach(function(event) {
-      topology.on(event, function(object1, object2) {
-        collectedEvents.push({
-          event: event,
-          object1: object1,
-          object2: object2
-        });
-      });
-    });
-  }
-
-  return collectedEvents;
-}
-
-//
-// Clear out all event
-//
-function clearAllEvents(topology) {
-  events.forEach(function(event) {
-    topology.removeAllListeners(event);
-  });
-}
-
-//
-// Replay any events due to single server connection switching to Mongos
-//
-function replayEvents(self, events) {
-  for (var i = 0; i < events.length; i++) {
-    self.emit(events[i].event, events[i].object1, events[i].object2);
-  }
-}
-
-function relayEvents(self, topology) {
-  var events = [
-    'serverOpening',
-    'serverDescriptionChanged',
-    'serverHeartbeatStarted',
-    'serverHeartbeatSucceeded',
-    'serverHeartbeatFailed',
-    'serverClosed',
-    'topologyOpening',
-    'topologyClosed',
-    'topologyDescriptionChanged',
-    'commandStarted',
-    'commandSucceeded',
-    'commandFailed',
-    'joined',
-    'left',
-    'ping',
-    'ha'
-  ];
-
-  events.forEach(function(event) {
-    topology.on(event, function(object1, object2) {
-      self.emit(event, object1, object2);
-    });
-  });
-}
-
-function assignTopology(client, topology) {
-  client.topology = topology;
-  topology.s.sessionPool = new ServerSessionPool(topology.s.coreTopology);
-}
-
-function createServer(self, options, callback) {
-  // Pass in the promise library
-  options.promiseLibrary = self.s.promiseLibrary;
-
-  // Set default options
-  var servers = translateOptions(options);
-
-  // Propagate the events to the client
-  var collectedEvents = collectEvents(self, servers[0]);
-
-  // Connect to topology
-  servers[0].connect(options, function(err, topology) {
-    if (err) return callback(err);
-    // Clear out all the collected event listeners
-    clearAllEvents(servers[0]);
-
-    // Relay all the events
-    relayEvents(self, servers[0]);
-    // Add listeners
-    addListeners(self, servers[0]);
-    // Check if we are really speaking to a mongos
-    var ismaster = topology.lastIsMaster();
-
-    // Set the topology
-    assignTopology(self, topology);
-
-    // Do we actually have a mongos
-    if (ismaster && ismaster.msg === 'isdbgrid') {
-      // Destroy the current connection
-      topology.close();
-      // Create mongos connection instead
-      return createMongos(self, options, callback);
-    }
-
-    // Fire all the events
-    replayEvents(self, collectedEvents);
-    // Otherwise callback
-    callback(err, topology);
-  });
-}
-
-function createReplicaset(self, options, callback) {
-  // Pass in the promise library
-  options.promiseLibrary = self.s.promiseLibrary;
-
-  // Set default options
-  var servers = translateOptions(options);
-
-  // Create the topology
-  var topology = new ReplSet(servers, options);
-
-  // Add listeners
-  addListeners(self, topology);
-
-  // Propagate the events to the client
-  relayEvents(self, topology);
-
-  // Open the connection
-  topology.connect(options, function(err, topology) {
-    if (err) return callback(err);
-
-    assignTopology(self, topology);
-    callback(null, topology);
-  });
-}
-
-function createMongos(self, options, callback) {
-  // Pass in the promise library
-  options.promiseLibrary = self.s.promiseLibrary;
-
-  // Set default options
-  var servers = translateOptions(options);
-
-  // Create the topology
-  var topology = new Mongos(servers, options);
-
-  // Add listeners
-  addListeners(self, topology);
-
-  // Propagate the events to the client
-  relayEvents(self, topology);
-
-  // Open the connection
-  topology.connect(options, function(err, topology) {
-    if (err) return callback(err);
-
-    assignTopology(self, topology);
-    callback(null, topology);
-  });
-}
-
-function createListener(self, event) {
-  return function(v1, v2) {
-    if (event === 'open' || event === 'fullsetup' || event === 'all' || event === 'reconnect') {
-      return self.emit(event, self);
-    }
-
-    self.emit(event, v1, v2);
-  };
-}
-
-function addListeners(self, topology) {
-  topology.on('authenticated', createListener(self, 'authenticated'));
-  topology.on('error', createListener(self, 'error'));
-  topology.on('timeout', createListener(self, 'timeout'));
-  topology.on('close', createListener(self, 'close'));
-  topology.on('parseError', createListener(self, 'parseError'));
-  topology.once('open', createListener(self, 'open'));
-  topology.once('fullsetup', createListener(self, 'fullsetup'));
-  topology.once('all', createListener(self, 'all'));
-  topology.on('reconnect', createListener(self, 'reconnect'));
-}
-
-function connectHandler(client, options, callback) {
-  return function(err, topology) {
-    if (err) {
-      return process.nextTick(function() {
-        try {
-          callback(err, null);
-        } catch (err) {
-          if (topology) topology.close();
-          throw err;
-        }
-      });
-    }
-
-    // No authentication just reconnect
-    if (!options.auth) {
-      return process.nextTick(function() {
-        try {
-          callback(err, topology);
-        } catch (err) {
-          if (topology) topology.close();
-          throw err;
-        }
-      });
-    }
-
-    // Authenticate
-    authenticate(client, options.user, options.password, options, function(err, success) {
-      if (success) {
-        process.nextTick(function() {
-          try {
-            callback(null, topology);
-          } catch (err) {
-            if (topology) topology.close();
-            throw err;
-          }
-        });
-      } else {
-        if (topology) topology.close();
-        process.nextTick(function() {
-          try {
-            callback(err ? err : new Error('Could not authenticate user ' + options.auth[0]), null);
-          } catch (err) {
-            if (topology) topology.close();
-            throw err;
-          }
-        });
-      }
-    });
-  };
-}
-
-/*
- * Connect using MongoClient
- */
-var connect = function(self, url, options, callback) {
-  options = options || {};
-  options = shallowClone(options);
-
-  // If callback is null throw an exception
-  if (callback == null) {
-    throw new Error('no callback function provided');
-  }
-
-  // Get a logger for MongoClient
-  var logger = Logger('MongoClient', options);
-
-  // Did we pass in a Server/ReplSet/Mongos
-  if (url instanceof Server || url instanceof ReplSet || url instanceof Mongos) {
-    // Set the topology
-    assignTopology(self, url);
-
-    // Add listeners
-    addListeners(self, url);
-
-    // Propagate the events to the client
-    relayEvents(self, url);
-
-    // Connect
-    return url.connect(
-      options,
-      connectHandler(self, options, function(err, topology) {
-        if (err) return connectCallback(err, topology);
-        if (options.user || options.password || options.authMechanism) {
-          return authenticate(self, options.user, options.password, options, function(err) {
-            if (err) return connectCallback(err, topology);
-            connectCallback(err, topology);
-          });
-        }
-
-        connectCallback(err, topology);
-      })
-    );
-  }
-
-  const parseFn = options.useNewUrlParser ? parse : legacyParse;
-  const transform = options.useNewUrlParser ? transformUrlOptions : legacyTransformUrlOptions;
-
-  parseFn(url, options, function(err, _object) {
-    // Do not attempt to connect if parsing error
-    if (err) return callback(err);
-
-    // Flatten
-    const object = transform(_object);
-
-    // Parse the string
-    const _finalOptions = createUnifiedOptions(object, options);
-
-    // Check if we have connection and socket timeout set
-    if (_finalOptions.socketTimeoutMS == null) _finalOptions.socketTimeoutMS = 360000;
-    if (_finalOptions.connectTimeoutMS == null) _finalOptions.connectTimeoutMS = 30000;
-
-    if (_finalOptions.db_options && _finalOptions.db_options.auth) {
-      delete _finalOptions.db_options.auth;
-    }
-
-    // Store the merged options object
-    self.s.options = _finalOptions;
-
-    // Failure modes
-    if (object.servers.length === 0) {
-      return callback(new Error('connection string must contain at least one seed host'));
-    }
-
-    // Do we have a replicaset then skip discovery and go straight to connectivity
-    if (_finalOptions.replicaSet || _finalOptions.rs_name) {
-      return createReplicaset(
-        self,
-        _finalOptions,
-        connectHandler(self, _finalOptions, connectCallback)
-      );
-    } else if (object.servers.length > 1) {
-      return createMongos(
-        self,
-        _finalOptions,
-        connectHandler(self, _finalOptions, connectCallback)
-      );
-    } else {
-      return createServer(
-        self,
-        _finalOptions,
-        connectHandler(self, _finalOptions, connectCallback)
-      );
-    }
-  });
-
-  function connectCallback(err, topology) {
-    if (err && err.message === 'no mongos proxies found in seed list') {
-      if (logger.isWarn()) {
-        logger.warn(
-          f(
-            'seed list contains no mongos proxies, replicaset connections requires the parameter replicaSet to be supplied in the URI or options object, mongodb://server:port/db?replicaSet=name'
-          )
-        );
-      }
-
-      // Return a more specific error message for MongoClient.connect
-      return callback(
-        new MongoError(
-          'seed list contains no mongos proxies, replicaset connections requires the parameter replicaSet to be supplied in the URI or options object, mongodb://server:port/db?replicaSet=name'
-        )
-      );
-    }
-
-    // Return the error and db instance
-    callback(err, topology);
-  }
 };
 
 module.exports = MongoClient;
